feat(nlu-server): model transfer is available without the need of an external server (#193)
* chore: rm model transfer server

* feat(nlu-server): model transfer is available without the need of an external server
