--- packages/mt-server/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@botpress/mt-server",
-  "description": "Model Transfer Server",
-  "version": "0.0.1",
-  "author": "Botpress, Inc.",
-  "license": "AGPL-3.0",
-  "scripts": {
-    "build": "tsc --build",
-    "clean": "rimraf ./dist && rimraf ./node_modules"
-  },
-  "dependencies": {
-    "@botpress/logger": "*",
-    "@botpress/nlu-client": "*",
-    "body-parser": "^1.18.3",
-    "bluebird": "^3.7.2",
-    "bytes": "^3.1.0",
-    "ms": "^2.1.1",
-    "moment": "^2.24.0",
-    "lodash": "^4.17.19",
-    "express": "^4.16.4",
-    "express-rate-limit": "^5.5.0",
-    "nanoid": "^3.1.23",
-    "yn": "^4.0.0"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/lodash": "^4.14.116",
-    "@types/node": "^16.11.10",
-    "@types/nanoid": "^3.0.0",
-    "typescript": "^4.5.2"
-  },
-  "types": "./dist/index.d.ts",
-  "main": "./dist/index.js"
-}
