--- packages/mt-server/src/error.ts
@@ -1,5 +1,0 @@
-export class ModelTransferError extends Error {
-  constructor(public readonly httpCode: number, msg: string) {
-    super(msg)
-  }
-}
