--- packages/mt-server/src/require-json.ts
@@ -1,6 +1,0 @@
-export const requireJSON = <T>(filePath: string): T | undefined => {
-  try {
-    const fileContent = require(filePath)
-    return fileContent
-  } catch (err) {}
-}
