--- packages/mt-server/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.packages.json",
-  "references": [{ "path": "../logger" }, { "path": "../nlu-client" }],
-  "compilerOptions": {
-    "outDir": "./dist" /* Redirect output structure to the directory. */,
-    "rootDir": "./src" /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
-    "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */,
-    "baseUrl": ".",
-    "composite": true,
-    "typeRoots": ["./node_modules/@types", "../../node_modules/@types"]
-  }
-}
