+++ packages/nlu-cli/src/parameters/index.ts
@@ -1,4 +1,3 @@
 export { parameters as nluServerParameters } from './nlu-server'
 export { parameters as langServerParameters } from './lang-server'
 export { parameters as langDownloadParameters } from './lang-download'
-export { parameters as modelTransferParameters } from './mt-server'
