+++ packages/nlu-cli/tsconfig.json
@@ -1,11 +1,6 @@
 {
   "extends": "../../tsconfig.packages.json",
-  "references": [
-    { "path": "../logger" },
-    { "path": "../nlu-server" },
-    { "path": "../lang-server" },
-    { "path": "../mt-server" }
-  ],
+  "references": [{ "path": "../logger" }, { "path": "../nlu-server" }, { "path": "../lang-server" }],
   "compilerOptions": {
     "outDir": "./dist" /* Redirect output structure to the directory. */,
     "rootDir": "./src" /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
