+++ packages/nlu-client/src/http-call.ts
@@ -1,0 +1,5 @@
+export type HTTPVerb = 'GET' | 'POST' | 'PUT' | 'DELETE'
+export type HTTPCall<V extends HTTPVerb> = {
+  verb: V
+  ressource: string
+}
