+++ packages/nlu-client/src/index.ts
@@ -1,8 +1,6 @@
 export { NLUClient as Client } from './nlu-client'
-export { ModelTransferClient as ModelClient } from './model-client'
 
 export * as http from './typings/http'
-export * as models from './typings/model'
 export * from './typings/training'
 export * from './typings/prediction'
 export * from './typings/linting'
