+++ packages/nlu-client/src/nlu-client.ts
@@ -1,6 +1,9 @@
 import axios, { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios'
 
 import _ from 'lodash'
+import { appIdHeader } from './app-id'
+import { ClientResponseError } from './error'
+import { HTTPCall, HTTPVerb } from './http-call'
 import { ModelTransferClient } from './model-client'
 import {
   TrainResponseBody,
@@ -18,27 +21,19 @@ import {
   ListTrainingsResponseBody,
   LintRequestBody,
   LintResponseBody,
-  LintProgressResponseBody,
-  DownloadModelResponseBody,
-  UploadModelRequestBody
+  LintProgressResponseBody
 } from './typings/http'
 import { IssueComputationSpeed } from './typings/linting'
-import { validateResponse, HTTPCall, HTTPVerb, ClientResponseError } from './validation'
-
-const DEFAULT_CONFIG: AxiosRequestConfig = {
-  validateStatus: () => true // return error datastructure instead of throwing
-}
+import { validateResponse } from './validation'
 
 export class NLUClient {
   protected _axios: AxiosInstance
-  public readonly modelTransfer: ModelTransferClient
+  public readonly modelWeights: ModelTransferClient
 
   constructor(config: AxiosRequestConfig & { baseURL: string }) {
-    config = { ...DEFAULT_CONFIG, ...config }
+    config = { ...config, validateStatus: () => true }
     this._axios = axios.create(config)
-
-    const modelTransferURL = `${config.baseURL}/modelweights`
-    this.modelTransfer = new ModelTransferClient({ baseURL: modelTransferURL })
+    this.modelWeights = new ModelTransferClient(config)
   }
 
   public get axios() {
@@ -52,25 +47,8 @@ export class NLUClient {
     return validateResponse<InfoResponseBody>(call, res)
   }
 
-  public async getModelWeightsInfo(appId: string, modelId: string): Promise<DownloadModelResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
-    const ressource = `model/${modelId}`
-    const call: HTTPCall<'GET'> = { verb: 'GET', ressource }
-    const res = await this._get(call, { headers })
-    return validateResponse<DownloadModelResponseBody>(call, res)
-  }
-
-  public async setModelWeightsInfo(appId: string, ressourceUUID: string): Promise<SuccessReponse | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
-    const ressource = 'model'
-    const call: HTTPCall<'POST'> = { verb: 'POST', ressource }
-    const body: UploadModelRequestBody = { ressourceUUID }
-    const res = await this._post(call, body, { headers })
-    return validateResponse<SuccessReponse>(call, res)
-  }
-
   public async startTraining(appId: string, body: TrainRequestBody): Promise<TrainResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = 'train'
     const call: HTTPCall<'POST'> = { verb: 'POST', ressource }
     const res = await this._post(call, body, { headers })
@@ -81,7 +59,7 @@ export class NLUClient {
    * @experimental still subject to breaking changes
    */
   public async startLinting(appId: string, body: LintRequestBody): Promise<LintResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = 'lint'
     const call: HTTPCall<'POST'> = { verb: 'POST', ressource }
     const res = await this._post(call, body, { headers })
@@ -89,7 +67,7 @@ export class NLUClient {
   }
 
   public async listTrainings(appId: string, lang?: string): Promise<ListTrainingsResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = 'train'
     const call: HTTPCall<'GET'> = { verb: 'GET', ressource }
     const params = lang && { lang }
@@ -98,7 +76,7 @@ export class NLUClient {
   }
 
   public async getTrainingStatus(appId: string, modelId: string): Promise<TrainProgressResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = `train/${modelId}`
     const call: HTTPCall<'GET'> = { verb: 'GET', ressource }
     const res = await this._get(call, { headers })
@@ -113,7 +91,7 @@ export class NLUClient {
     modelId: string,
     speed: IssueComputationSpeed
   ): Promise<LintProgressResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = `lint/${modelId}/${speed}`
     const call: HTTPCall<'GET'> = { verb: 'GET', ressource }
     const res = await this._get(call, { headers })
@@ -121,7 +99,7 @@ export class NLUClient {
   }
 
   public async cancelTraining(appId: string, modelId: string): Promise<SuccessReponse | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = `train/${modelId}/cancel`
     const call: HTTPCall<'POST'> = { verb: 'POST', ressource }
     const res = await this._post(call, {}, { headers })
@@ -136,7 +114,7 @@ export class NLUClient {
     modelId: string,
     speed: IssueComputationSpeed
   ): Promise<SuccessReponse | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = `lint/${modelId}/${speed}/cancel`
     const call: HTTPCall<'POST'> = { verb: 'POST', ressource }
     const res = await this._post(call, {}, { headers })
@@ -144,7 +122,7 @@ export class NLUClient {
   }
 
   public async listModels(appId: string): Promise<ListModelsResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = 'models'
     const call: HTTPCall<'GET'> = { verb: 'GET', ressource }
     const res = await this._get(call, { headers })
@@ -152,7 +130,7 @@ export class NLUClient {
   }
 
   public async pruneModels(appId: string): Promise<PruneModelsResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = 'models/prune'
     const call: HTTPCall<'POST'> = { verb: 'POST', ressource }
     const res = await this._post(call, {}, { headers })
@@ -163,7 +141,7 @@ export class NLUClient {
     appId: string,
     body: DetectLangRequestBody
   ): Promise<DetectLangResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = 'detect-lang'
     const call: HTTPCall<'POST'> = { verb: 'POST', ressource }
     const res = await this._post(call, body, { headers })
@@ -175,7 +153,7 @@ export class NLUClient {
     modelId: string,
     body: PredictRequestBody
   ): Promise<PredictResponseBody | ErrorResponse> {
-    const headers = this._appIdHeader(appId)
+    const headers = appIdHeader(appId)
     const ressource = `predict/${modelId}`
     const call: HTTPCall<'POST'> = { verb: 'POST', ressource }
     const res = await this._post(call, body, { headers })
@@ -212,11 +190,5 @@ export class NLUClient {
     const err = thrown instanceof Error ? thrown : new Error(`${thrown}`)
     const httpStatus = -1
     return new ClientResponseError(call, httpStatus, err.message)
-  }
-
-  private _appIdHeader = (appId: string) => {
-    return {
-      'X-App-Id': appId
-    }
   }
 }
