--- packages/nlu-client/src/typings/model.ts
@@ -1,8 +1,0 @@
-export type UploadModelResponse = {
-  uuid: string
-  ttl: number
-}
-
-export type ModelServerInfo = {
-  version: string
-}
