+++ packages/nlu-server/src/application/index.ts
@@ -7,10 +7,8 @@ import {
   TrainingStatus,
   LintingState,
   IssueComputationSpeed,
-  ModelClient,
   ModelTransferInfo
 } from '@botpress/nlu-client'
-import { UploadModelResponse } from '@botpress/nlu-client/src/typings/model'
 import { Engine, ModelId, modelIdService, errors as engineErrors } from '@botpress/nlu-engine'
 import axios from 'axios'
 import Bluebird from 'bluebird'
@@ -27,16 +25,18 @@ import {
   DucklingCommError,
   InvalidModelSpecError,
   DatasetValidationError,
-  LintingNotFoundError,
-  ModelTransferError
+  LintingNotFoundError
 } from './errors'
 import { LintingQueue } from './linting-queue'
 import { deserializeModel } from './serialize-model'
 import { TrainingQueue } from './training-queue'
 
+type AppOptions = {
+  modelTransferEnabled: boolean
+}
+
 export class Application extends ApplicationObserver {
   private _logger: Logger
-  private _modelTransfer: ModelTransferInfo = { enabled: false }
 
   constructor(
     private _modelRepo: ModelRepository,
@@ -47,7 +47,7 @@ export class Application extends ApplicationObserver {
     private _engine: Engine,
     private _serverVersion: string,
     baseLogger: Logger,
-    private _modelClient?: ModelClient
+    private _opts: Partial<AppOptions> = {}
   ) {
     super()
     this._logger = baseLogger.sub('app')
@@ -60,11 +60,6 @@ export class Application extends ApplicationObserver {
     await this._lintingRepo.initialize()
     await this._lintingQueue.initialize()
     this._trainingQueue.addListener(this._listenTrainingUpdates)
-
-    if (this._modelClient) {
-      const { version } = await this._modelClient.getInfo()
-      this._modelTransfer = { enabled: true, version }
-    }
   }
 
   public async teardown() {
@@ -82,11 +77,13 @@ export class Application extends ApplicationObserver {
     const specs = this._engine.getSpecifications()
     const languages = this._engine.getLanguages()
     const version = this._serverVersion
-    return { specs, languages, version, modelTransfer: this._modelTransfer }
+
+    const { modelTransferEnabled } = this._opts
+    return { specs, languages, version, modelTransfer: { enabled: !!modelTransferEnabled } }
   }
 
-  public async prepareModelForDownload(appId: string, modelId: ModelId): Promise<UploadModelResponse> {
-    if (!this._modelClient) {
+  public async getModelWeights(appId: string, modelId: ModelId): Promise<Buffer> {
+    if (!this._opts.modelTransferEnabled) {
       throw new ModelTransferDisabled()
     }
 
@@ -95,27 +92,16 @@ export class Application extends ApplicationObserver {
       throw new ModelDoesNotExistError(appId, modelId)
     }
 
-    return this._modelClient.uploadModelWeights(modelWeights)
+    return modelWeights
   }
 
-  public async recoverUploadedModel(appId: string, uuid: string) {
-    if (!this._modelClient) {
+  public async setModelWeights(appId: string, modelWeights: Buffer) {
+    if (!this._opts.modelTransferEnabled) {
       throw new ModelTransferDisabled()
     }
-
-    try {
-      const modelWeights = await this._modelClient.downloadModelWeights(uuid, { responseType: 'arraybuffer' })
-
-      // TODO: validate model weights
-      const model = await deserializeModel(modelWeights)
-
-      return this._modelRepo.saveModel(appId, model.id, modelWeights)
-    } catch (err) {
-      if (axios.isAxiosError(err)) {
-        throw new ModelTransferError(err)
-      }
-      throw err
-    }
+    // TODO: validate model weights
+    const model = await deserializeModel(modelWeights)
+    return this._modelRepo.saveModel(appId, model.id, modelWeights)
   }
 
   public async getModels(appId: string): Promise<ModelId[]> {
