+++ packages/nlu-server/src/bootstrap/make-application.ts
@@ -1,5 +1,4 @@
 import { Logger } from '@botpress/logger'
-import { ModelClient } from '@botpress/nlu-client'
 import { Engine } from '@botpress/nlu-engine'
 import Knex from 'knex'
 import { Application } from '../application'
@@ -79,8 +78,7 @@ export const makeApplication = async (
     options
   )
 
-  const modelClient = options.modelTransferURL ? new ModelClient({ baseURL: options.modelTransferURL }) : undefined
-
+  const { modelTransferEnabled } = options
   const application = new Application(
     modelRepo,
     trainRepo,
@@ -90,7 +88,7 @@ export const makeApplication = async (
     engine,
     serverVersion,
     baseLogger,
-    modelClient
+    { modelTransferEnabled }
   )
   await application.initialize()
   return application
